pyproject.toml
[project]
name = "quickgrade"
version = "0.1.0"
description = "GitHub & Taiga analytics dashboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Django
    "django>=5.1",
    "django-allauth>=65.0",
    "django-environ>=0.11",

    # Database
    "psycopg[binary]>=3.2",

    # Async HTTP
    "httpx>=0.28",

    # Task Queue
    "celery>=5.4",
    "redis>=5.2",

    # API Clients
    "PyGithub>=2.5",
    "python-taiga>=1.3",

    # Data Processing
    "pandas>=2.2",
    "openpyxl>=3.1",

    # Code Analysis
    "lizard>=1.17",
    "complexipy>=0.4",

    # Git Operations
    "gitpython>=3.1",

    # CLI
    "rich>=13.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "quickgrade.settings"
python_files = ["test_*.py", "*_test.py"]
